机构-增删改查

+++ src/main/java/com/hd/home_disabled/repository/UserRepository.java
@@ -3,5 +3,12 @@ package com.hd.home_disabled.repository;
 import com.hd.home_disabled.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User,Long> {
+/**
+ * @ClassName UserRepository
+ * @Description TODO
+ * @Author pyt
+ * @Date 2019/7/5 11:00
+ * @Version
+ */
+public interface UserRepository extends JpaRepository<User, Integer> {
 }
